+++ src/xtd.forms.native.cocoa/src/xtd/forms/native/cocoa/message_box.mm
@@ -10,8 +10,8 @@ using namespace xtd;
 using namespace xtd::forms::native;
 
 namespace {
-  constexpr int32_t MB_BUTTONS_MASK = 0x7;
-  constexpr int32_t MB_ICON_MASK = 0x70;
+  constexpr int_least32_t MB_BUTTONS_MASK = 0x7;
+  constexpr int_least32_t MB_ICON_MASK = 0x70;
   
   static void message_box_add_buttons_ok(NSAlert* alert) {
     [alert addButtonWithTitle:@"OK"];
@@ -44,16 +44,16 @@ namespace {
     [alert addButtonWithTitle:@"Cancel"];
   }
   
-  static int32_t message_box_show_modal_ok(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_ok(NSAlert* alert) {
     [alert runModal];
     return IDOK;
   }
   
-  static int32_t message_box_show_modal_ok_cancel(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_ok_cancel(NSAlert* alert) {
     return [alert runModal] == NSAlertFirstButtonReturn ? IDOK : IDCANCEL;
   }
   
-  static int32_t message_box_show_modal_abort_retry_ignore(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_abort_retry_ignore(NSAlert* alert) {
     NSModalResponse result = [alert runModal];
     if (result == NSAlertFirstButtonReturn)
       return IDABORT;
@@ -62,7 +62,7 @@ namespace {
     return IDIGNORE;
   }
   
-  static int32_t message_box_show_modal_yes_no_cancel(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_yes_no_cancel(NSAlert* alert) {
     NSModalResponse result = [alert runModal];
     if (result == NSAlertFirstButtonReturn)
       return IDYES;
@@ -71,15 +71,15 @@ namespace {
     return IDCANCEL;
   }
   
-  static int32_t message_box_show_modal_yes_no(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_yes_no(NSAlert* alert) {
     return [alert runModal] == NSAlertFirstButtonReturn ? IDYES : IDNO;
   }
   
-  static int32_t message_box_show_modal_retry_cancel(NSAlert* alert) {
+  static int_least32_t message_box_show_modal_retry_cancel(NSAlert* alert) {
     return [alert runModal] == NSAlertFirstButtonReturn ? IDRETRY : IDCANCEL;
   }
   
-  static void add_buttons(uint32_t style, NSAlert* alert) {
+  static void add_buttons(uint_least32_t style, NSAlert* alert) {
     if ((style & MB_BUTTONS_MASK) == MB_RETRYCANCEL) message_box_add_buttons_retry_cancel(alert);
     else if ((style & MB_BUTTONS_MASK) == MB_YESNOCANCEL) message_box_add_buttons_yes_no_cancel(alert);
     else if ((style & MB_BUTTONS_MASK) == MB_YESNO) message_box_add_buttons_yes_no(alert);
@@ -88,7 +88,7 @@ namespace {
     else message_box_add_buttons_ok(alert);
   }
   
-  static NSAlertStyle convert_to_icon(int32_t style) {
+  static NSAlertStyle convert_to_icon(int_least32_t style) {
     if ((style & MB_ICON_MASK) == MB_ICONINFORMATION) return NSAlertStyleInformational;
     if ((style & MB_ICON_MASK) == MB_ICONEXCLAMATION) return NSAlertStyleCritical;
     if ((style & MB_ICON_MASK) == MB_ICONQUESTION) return NSAlertStyleInformational;
@@ -96,7 +96,7 @@ namespace {
     return NSAlertStyleWarning;
   }
   
-  static int32_t show_modal(int32_t style, NSAlert* alert) {
+  static int_least32_t show_modal(int_least32_t style, NSAlert* alert) {
     if ((style & MB_BUTTONS_MASK) == MB_RETRYCANCEL) return message_box_show_modal_retry_cancel(alert);
     if ((style & MB_BUTTONS_MASK) == MB_YESNOCANCEL) return message_box_show_modal_yes_no_cancel(alert);
     if ((style & MB_BUTTONS_MASK) == MB_YESNO) return message_box_show_modal_yes_no(alert);
@@ -106,7 +106,7 @@ namespace {
   }
 }
 
-int32_t message_box::show(intptr_t control, const std::string& text, const std::string& caption, uint32_t style, bool displayHelpButton) {
+int_least32_t message_box::show(intmax_t control, const std::string& text, const std::string& caption, uint_least32_t style, bool displayHelpButton) {
   @autoreleasepool {
     NSAlert* alert = [[NSAlert alloc] init];
     add_buttons(style, alert);
